#572 removed gulp build system to just npm scripts, no minified version now. updated rollup and babel (and presets)
