--- gulpfile.babel.js
@@ -1,5 +1,0 @@
-import gulp from 'gulp';
-
-import './gulp/build';
-import './gulp/test';
-import './gulp/watch';
