+++ package.json
@@ -16,8 +16,8 @@
     "url": "https://github.com/PaulUithol/Backbone-relational.git"
   },
   "scripts": {
-    "test": "node_modules/.bin/karma start",
-    "build": "node_modules/.bin/gulp build"
+    "test": "karma start",
+    "build": "rollup -c"
   },
   "author": "Paul Uithol <https://github.com/PaulUithol>",
   "contributors": "Listed at <https://github.com/PaulUithol/Backbone-relational/contributors>",
@@ -28,21 +28,12 @@
   "license": "MIT",
   "version": "0.10.0",
   "devDependencies": {
-    "babel-core": "^6.8.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-es2015-rollup": "^1.1.1",
-    "babel-register": "^6.8.0",
+    "babel-plugin-external-helpers": "^6.8.0",
+    "babel-polyfill": "^6.9.1",
+    "babel-preset-es2015": "^6.16.0",
     "babelify": "^7.3.0",
     "backbone": "1.2.3",
     "browserify": "^13.0.0",
-    "gulp-file": "^0.3.0",
-    "gulp-filter": "^4.0.0",
-    "gulp-plumber": "^1.1.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-rollup": "^1.9.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-uglify": "^1.5.3",
-    "jasmine-core": "^2.4.1",
     "jquery": "^2.2.1",
     "karma": "^0.13.22",
     "karma-browserify": "^5.0.2",
@@ -50,11 +41,9 @@
     "karma-phantomjs-launcher": "^1.0.0",
     "karma-qunit": "^0.1.9",
     "phantomjs-prebuilt": "^2.1.5",
-    "rollup": "^0.26.3",
-    "rollup-plugin-babel": "^2.4.0",
+    "rollup": "^0.36.3",
+    "rollup-plugin-babel": "^2.6.1",
     "rollup-plugin-json": "^2.0.0"
   },
-  "dependencies": {
-    "babel-polyfill": "^6.9.1"
-  }
+  "dependencies": {}
 }
